starboard: Initialize Starboard FeatureList singletons in unit and browser tests
Fixes uninitialized Starboard FeatureList singletons when running tests on Starboard and Ozone Starboard platforms.

1. starboard/features_test_util: Updates InitializeStarboardFeatureListWithDefaults() to query SbSystemGetExtension(kStarboardExtensionFeaturesName) and invoke the extension's InitializeStarboardFeatures function if available. This ensures Starboard feature flags are initialized in the host process for Evergreen/modular test builds (e.g. nplb_loader).
2. starboard/shared/starboard: Refactors StarboardTestEnvironment::SetUp() to use features::InitializeStarboardFeatureListWithDefaults().
3. ui/ozone/platform/starboard: Ensures Starboard features are initialized prior to PlatformWindowStarboard creation in ozone_unittests.
4. ui/wm: Initializes Starboard features in WMTestSuite::Initialize() for wm_unittests on hermetic builds.
5. starboard/command_line_preprocessor: Refactors ConvertDeprecatedSwitches into command_line_preprocessor module.
6. content/shell: Overrides BasicStartupComplete() in ContentBrowserTestShellMainDelegate to initialize Starboard features prior to browser test window creation.

TAG=agy
CONV=8cfe5487-b42a-485c-82e1-0911fd63cdc9
