+++ cobalt/testing/browser_tests/BUILD.gn
@@ -101,6 +101,7 @@ static_library("cobalt_browsertests_test_runner") {
     "//base/test:test_config",
     "//build:chromeos_buildflags",
     "//cobalt/browser:browser",
+    "//cobalt/common/features",
     "//cobalt/shell:cobalt_shell_app",
     "//components/network_session_configurator/common:common",
     "//components/services/storage:storage",
