+++ starboard/extension/BUILD.gn
@@ -34,8 +34,7 @@ source_set("extension_test") {
     "//testing/gmock",
     "//testing/gtest",
   ]
-  if (defined(sb_is_modular) && sb_is_modular && defined(cobalt_toolchain) &&
-      current_toolchain == cobalt_toolchain) {
+  if (sb_is_modular && current_toolchain == cobalt_toolchain) {
     deps += cobalt_platform_dependencies
   }
 }
